move internal/types and internal/pkg to pkg/lib
