+++ internal/util/pkgutil/pkgutil.go
@@ -22,9 +22,9 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/kptdev/kpt/internal/pkg"
 	kptfilev1 "github.com/kptdev/kpt/pkg/api/kptfile/v1"
 	"github.com/kptdev/kpt/pkg/kptfile/kptfileutil"
+	"github.com/kptdev/kpt/pkg/lib/pkg"
 	"sigs.k8s.io/kustomize/kyaml/copyutil"
 	"sigs.k8s.io/kustomize/kyaml/filesys"
 	"sigs.k8s.io/kustomize/kyaml/kio"
