+++ internal/util/pkgutil/pkgutil_test.go
@@ -20,11 +20,11 @@ import (
 	"sort"
 	"testing"
 
-	"github.com/kptdev/kpt/internal/pkg"
 	"github.com/kptdev/kpt/internal/testutil"
 	"github.com/kptdev/kpt/internal/testutil/pkgbuilder"
 	"github.com/kptdev/kpt/internal/util/pkgutil"
 	kptfilev1 "github.com/kptdev/kpt/pkg/api/kptfile/v1"
+	"github.com/kptdev/kpt/pkg/lib/pkg"
 	"github.com/stretchr/testify/assert"
 )
