+++ internal/util/update/resource-merge.go
@@ -21,10 +21,9 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/kptdev/kpt/pkg/lib/types"
 	"sigs.k8s.io/kustomize/kyaml/pathutil"
 
-	"github.com/kptdev/kpt/internal/pkg"
-	"github.com/kptdev/kpt/internal/types"
 	pkgdiff "github.com/kptdev/kpt/internal/util/diff"
 	"github.com/kptdev/kpt/internal/util/merge"
 	"github.com/kptdev/kpt/internal/util/pkgutil"
@@ -32,7 +31,8 @@ import (
 	kptfilev1 "github.com/kptdev/kpt/pkg/api/kptfile/v1"
 	"github.com/kptdev/kpt/pkg/kptfile/kptfileutil"
 	"github.com/kptdev/kpt/pkg/lib/errors"
-	updatetypes "github.com/kptdev/kpt/pkg/lib/update/updatetypes"
+	"github.com/kptdev/kpt/pkg/lib/pkg"
+	"github.com/kptdev/kpt/pkg/lib/update/updatetypes"
 	"github.com/kptdev/krm-functions-sdk/go/fn"
 	"gopkg.in/yaml.v2"
 	"k8s.io/klog/v2"
