+++ pkg/fn/runtime/runner_test.go
@@ -24,9 +24,9 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/kptdev/kpt/internal/types"
 	kptfilev1 "github.com/kptdev/kpt/pkg/api/kptfile/v1"
 	"github.com/kptdev/kpt/pkg/lib/runneroptions"
+	"github.com/kptdev/kpt/pkg/lib/types"
 	"github.com/kptdev/kpt/pkg/printer"
 	"github.com/stretchr/testify/assert"
 	"sigs.k8s.io/kustomize/kyaml/filesys"
