+++ pkg/kptfile/kptfileutil/util_test.go
@@ -1386,7 +1386,7 @@ pipeline:
 			assert.NoError(t, err)
 			originKf, err := DecodeKptfile(strings.NewReader(tc.origin))
 			assert.NoError(t, err)
-			err = merge(localKf, updatedKf, originKf)
+			err = MergeKptfiles(localKf, updatedKf, originKf)
 			if tc.err == nil {
 				if !assert.NoError(t, err) {
 					t.FailNow()
