+++ pkg/lib/errors/resolver/live.go
@@ -20,7 +20,7 @@ import (
 	"strings"
 
 	initialization "github.com/kptdev/kpt/commands/live/init"
-	"github.com/kptdev/kpt/internal/pkg"
+	"github.com/kptdev/kpt/pkg/lib/pkg"
 	"github.com/kptdev/kpt/pkg/lib/util/cmdutil"
 	"github.com/kptdev/kpt/pkg/live"
 	"sigs.k8s.io/cli-utils/pkg/inventory"
