+++ pkg/lib/errors/resolver/pkg.go
@@ -18,10 +18,10 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/kptdev/kpt/internal/pkg"
 	kptfilev1 "github.com/kptdev/kpt/pkg/api/kptfile/v1"
 	"github.com/kptdev/kpt/pkg/kptfile/kptfileutil"
 	"github.com/kptdev/kpt/pkg/lib/errors"
+	"github.com/kptdev/kpt/pkg/lib/pkg"
 )
 
 //nolint:gochecknoinits
